Conflicts:
	components/tools/OmeroPy/test/gatewaytest/chgrp.py

+++ components/tools/OmeroPy/src/omero/gateway/__init__.py
@@ -25,7 +25,8 @@ import ConfigParser
 import omero
 import omero.clients
 from omero.util.decorators import timeit, TimeIt, setsessiongroup
-from omero.cmd import Chgrp
+from omero.cmd import Chgrp, DoAll
+from omero.api import Save
 from omero.callbacks import CmdCallbackI
 from omero.gateway.utils import ServiceOptsDict, GatewayConfig
 import omero.scripts as scripts
@@ -742,12 +743,15 @@ class BlitzObjectWrapper (object):
 
         # Using omero.cmd.Delete rather than deleteObjects since we need
         # spec/id pairs rather than spec+id_list as arguments
-        handle = self._conn.c.sf.submit(dcs, self._conn.SERVICE_OPTS)
-        callback = omero.callbacks.CmdCallbackI(self._conn.c, handle)
-        # Maximum wait time 5 seconds, will raise a LockTimeout if the
-        # delete has not finished by then.
-        callback.loop(10, 500)
-        self._obj.unloadAnnotationLinks()
+        if len(dcs):
+            doall = omero.cmd.DoAll()
+            doall.requests = dcs
+            handle = self._conn.c.sf.submit(doall, self._conn.SERVICE_OPTS)
+            try:
+                self._conn._waitOnCmd(handle)
+            finally:
+                handle.close()
+            self._obj.unloadAnnotationLinks()
 
     def removeAnnotations (self, ns):
         """
@@ -763,12 +767,13 @@ class BlitzObjectWrapper (object):
         for al in self._getAnnotationLinks(ns=ns):
             a = al.child
             ids.append(a.id.val)
-        handle = self._conn.deleteObjects('/Annotation', ids)
-        callback = omero.callbacks.CmdCallbackI(self._conn.c, handle)
-        # Maximum wait time 5 seconds, will raise a LockTimeout if the
-        # delete has not finished by then.
-        callback.loop(10, 500)
-        self._obj.unloadAnnotationLinks()        
+        if len(ids):
+            handle = self._conn.deleteObjects('/Annotation', ids)
+            try:
+                self._conn._waitOnCmd(handle)
+            finally:
+                handle.close()
+            self._obj.unloadAnnotationLinks()        
     
     # findAnnotations(self, ns=[])
     def getAnnotation (self, ns=None):
@@ -3155,16 +3160,16 @@ class _BlitzGateway (object):
         return prx
 
 
-    def chgrpObject(self, graph_spec, obj_id, group_id):
+    def chgrpObjects(self, graph_spec, obj_ids, group_id, container_id=None):
         """
-        Change the Group for a specified object using queue.
+        Change the Group for a specified objects using queue.
 
         @param graph_spec:      String to indicate the object type or graph
                                 specification. Examples include:
                                  * '/Image'
                                  * '/Project'   # will move contents too.
                                  * NB: Also supports 'Image' etc for convenience
-        @param obj_id:          ID for the object to move.
+        @param obj_ids:         IDs for the objects to move.
         @param group_id:        The group to move the data to.
         """
 
@@ -3172,9 +3177,31 @@ class _BlitzGateway (object):
             graph_spec = '/%s' % graph_spec
             logger.debug('chgrp Received object type, using "%s"' % graph_spec)
 
-        chgrp = omero.cmd.Chgrp(type=graph_spec, id=obj_id, options=None, grp=group_id)
+        # (link, child, parent)
+        parentLinkClasses = {"/Image": (omero.model.DatasetImageLinkI, omero.model.ImageI, omero.model.DatasetI),
+                        "/Dataset": (omero.model.ProjectDatasetLinkI, omero.model.DatasetI, omero.model.ProjectI),
+                        "/Plate": (omero.model.ScreenPlateLinkI, omero.model.PlateI, omero.model.ScreenI)}
+        da = DoAll()
+        requests = []
+        for obj_id in obj_ids:
+            obj_id = long(obj_id)
+            logger.debug('DoAll Chgrp: type: %s, id: %s, grp: %s' % (graph_spec, obj_id, group_id))
+            chgrp = omero.cmd.Chgrp(type=graph_spec, id=obj_id, options=None, grp=group_id)
+            requests.append(chgrp)
+            if container_id is not None and graph_spec in parentLinkClasses:
+                # get link class for graph_spec objects
+                link_klass = parentLinkClasses[graph_spec][0]
+                link = link_klass()
+                link.child = parentLinkClasses[graph_spec][1](obj_id, False)
+                link.parent = parentLinkClasses[graph_spec][2](container_id, False)
+                save = Save()
+                save.obj = link
+                requests.append(save)
 
-        prx = self.c.sf.submit(chgrp)
+        da.requests = requests
+        ctx = self.SERVICE_OPTS.copy()
+        ctx.setOmeroGroup(group_id)         # NB: For Save to work, we need to be in target group
+        prx = self.c.sf.submit(da, ctx)
         return prx
 
 
@@ -6599,7 +6626,12 @@ class _ImageWrapper (BlitzObjectWrapper):
         for chunk in ofw.getFileInChunks():
             outfile.write(chunk)
         outfile.close()
-        self._conn.deleteObjects('/OriginalFile', todel) # No error handling?
+        handle = self._conn.deleteObjects('/OriginalFile', todel)
+        try:
+            self._conn._waitOnCmd(handle)
+        finally:
+            handle.close()
+
         return os.path.splitext(f.name.val)[-1], f.mimetype.val
         
     def renderImage (self, z, t, compression=0.9):
@@ -7019,7 +7051,10 @@ class _ImageWrapper (BlitzObjectWrapper):
 
     def _deleteSettings(self):
         handle = self._conn.deleteObjects("/Image/Pixels/RenderingDef", [self.getId()])
-        self._conn._waitOnCmd(handle)
+        try:
+            self._conn._waitOnCmd(handle)
+        finally:
+            handle.close()
 
     def _collectRenderOptions (self):
         """
